refactor: 语音合一重构——群聊/私聊统一单脑语音管线
- 新增 voice 模块与 ai::tool_loop：一次 LLM 调用同时完成感知、决策与表达，纯文本响应即最终发言，空响应/finish=沉默
- 状态注入体验化：情绪/电量/节律/注意力转为第一人称体验句（conversation/context.rs 重写为 VoiceScene）
- 群聊上下文改为带名字与时间戳的场景记录，模型可做多视角社交推理
- 删除 planner/replyer/timing_gate/typo 模块与 runtime 脚手架，移除 satisficing、随机口头禅、感叹号手术、句子截断、错别字注入等后处理手术
- 配额改为先说话后扣账，新增 voice_max_rounds 与 silence_cooldown_secs 配置
- 同步 config.example.yaml 与 SYSTEM_OVERVIEW.md

+++ src/circadian.rs
@@ -157,7 +157,11 @@ pub fn get_circadian_context(rhythm: &CircadianRhythm) -> String {
         descriptors.push("今天很有耐心，可以听别人慢慢说");
     }
 
-    lines.push(format!("# 时间与状态\n{}。{}", time_desc, descriptors.join("。")));
+    lines.push(format!(
+        "# 时间与状态\n{}。{}",
+        time_desc,
+        descriptors.join("。")
+    ));
 
     lines.join("\n")
 }
@@ -165,7 +169,7 @@ pub fn get_circadian_context(rhythm: &CircadianRhythm) -> String {
 /// 检查当前是否在免打扰时段（结合 proactive 配置）
 pub fn is_quiet_hours() -> bool {
     let cfg = config::get();
-    let hour = crate::util::current_hour_cst() as u32;
+    let hour = crate::util::current_hour_cst();
     let quiet_start = cfg.proactive.quiet_start;
     let quiet_end = cfg.proactive.quiet_end;
